Refactor public API
The public API has been re-worked in order to make DelimitatedDataParser
easier to use and to test.

The dependencies on `TextReader` and `DataTable` have been changed from
constructor parameters to method parameters on the `Parse` and `Export`
methods. This makes mocking and dependency injection easier and also
allows a single instance of the `Parser` and `Exporter` classes to be
re-used for multiple operations.

Fixes #5 and #6.